requirements.txt
discord.py
TESTE